move test helper into sep. file
